Move parsed_number_format into lib/src
Update imports and exports to the new lib/src paths.
Re-export NumberEditingTextController from
lib/number_editing_controller.dart.
Update example/pubspec.lock dependency versions and bump Dart SDK lower
bound to >=3.9.0-0.
