+++ lib/number_editing_controller.dart
@@ -1,1 +1,1 @@
-export 'parsed_number_format/text_controller.dart';
+export 'src/text_controller.dart' show NumberEditingTextController;
