+++ lib/src/part_length.dart
